refactor: Ajustar visuais

+++ src/components/notes/View.vue
@@ -1,9 +1,9 @@
 <template>
-  <div class="w-full">
+  <div class="max-w-xl bg-gray-100 shadow-lg px-2 py-4 border border-gray-100">
     <div class="mb-4 block">
       <label class="block text-gray-700 text-sm font-bold mb-2" for="note-titulo">Título</label>
       <input
-        class="block shadow appearance-none border rounded py-2 px-4 text-gray-700 leading-tight"
+        class="w-full block shadow appearance-none border rounded py-2 px-4 text-gray-700 leading-tight"
         type="text"
         id="note-titulo"
         disabled
@@ -13,23 +13,23 @@
     <div class="mb-4 block">
       <label class="block text-gray-700 text-sm font-bold mb-2" for="note-titulo">Descricao</label>
       <textarea
-        class="block shadow appearance-none border rounded py-2 px-4 text-gray-700 leading-tight"
+        class="w-full block shadow appearance-none border rounded py-2 px-4 text-gray-700 leading-tight"
         type="text"
         id="note-desc"
         disabled
         v-bind:value="selectedNote.descricao"
       ></textarea>
     </div>
-    <div class="block space-x-1">
+    <div class="relative block h-12">
       <button
         @click="editNote"
         type="button"
-        class="bg-gray-500 hover:bg-gray-300 transition duration-500 rounded-lg px-4 py-2 w-32"
+        class="button absolute right-0 bg-gray-500 hover:bg-gray-300"
       >Editar</button>
       <button
         @click="fireDeleteNote"
         type="button"
-        class="bg-red-600 hover:bg-red-400 transition duration-500 rounded-lg px-4 py-2 w-32"
+        class="button absolute left-0 bg-red-600 hover:bg-red-400"
       >Deletar</button>
     </div>
   </div>
@@ -47,8 +47,10 @@ export default defineComponent({
       ctx.emit('note-edit-enabled')
     }
     const fireDeleteNote = () => {
-      deleteNote(selectedNote.value)
-      ctx.emit('note-deleted')
+      if (confirm('Deseja deletar a anotação?')) {
+        deleteNote(selectedNote.value)
+        ctx.emit('note-deleted')
+      }
     }
     return {
       selectedNote,
@@ -59,5 +61,10 @@ export default defineComponent({
 })
 </script>
 
-<style>
+<style scoped lang="postcss">
+.button {
+  @apply font-semibold;
+  @apply rounded-md px-4 py-2 w-32 h-12;
+  @apply transition duration-300;
+}
 </style>
